// src/types/apps/menuTypes.ts
export type MenuesType = {
  id: number
  title: string
  availbility?: string
  availability: string
  sku: string
  brand: string
  rating: string
  price: string
  condition: string
  description: string
  link: string
  image_link: string
  restaurant: number
  name?: string
  business: number
  status: string
  availability_circle_radius: number
  menu_number: string
  flow: boolean
  template: boolean
}